RavenDB-11310 - Move InstallSnapshot, ReadSnapshot and ReadInstallSnapshotAndIgnoreContent from Follower to FollowerReadAndCommitSnapshotCommand
